requirements.txt
boto3>=1.38.18
cachetools>=5.5.2
discord.py>=2.3.2
email-validator>=1.1.3 # Or a more recent compatible version, check PyPI
Flask>=3.0.0
Flask-APScheduler>=1.13.1
Flask-Babel>=4.0.0
Flask-HTMX>=0.4.0
Flask-Login>=0.6.3
Flask-Migrate>=4.1.0
Flask-Session>=0.8.0
Flask-SQLAlchemy>=3.1.1
Flask-WTF>=1.2.2
gunicorn>=22.0.0
Markdown>=3.6
packaging>=24.0
plexapi>=4.17.0
python-dotenv>=1.1.0
requests>=2.32.3
SQLAlchemy>=2.0.41
# WTForms is a dep of Flask-WTF